Avoid clang 15 crash by using correct array size
Clang 15 crashes when compiling this code because the size of the array
does not match the initializer. This is fixed in clang 16, but

  Apple clang version 14.0.3 (clang-1403.0.22.14.1)

which is currently shipped in Ventura is still affected by this.

See https://github.com/llvm/llvm-project/issues/56016 for the upstream
report of this problem.

Signed-off-by: Clemens Lang <cal@macports.org>
(cherry picked from commit 759b921e3b5b2422787c9381140c27d065c6bd48)